+++ src/GitHUD/Terminal/Types.hs
@@ -6,4 +6,4 @@ module GitHUD.Terminal.Types (
 
 data Color = Black | Red | Green | Yellow | Blue | Magenta | Cyan | White | NoColor deriving (Eq, Show, Read)
 data ColorIntensity = Dull | Vivid deriving (Eq, Show, Read)
-data Shell = ZSH | BASH | TMUX | Other deriving (Eq, Show)
+data Shell = ZSH | BASH | TMUX | NONE | Other deriving (Eq, Show)